hopefully this helps

+++ src/c3nav/routing/locator.py
@@ -147,7 +147,7 @@ class LocatorSpace:
         )
         if not acceptable_points:
             return None, None
-        scores = np.sum((self.levels[np.array(acceptable_points).reshape((-1, 1)),
+        scores = np.sum((self.levels[np.array(acceptable_points, dtype=np.uint32).reshape((-1, 1)),
                                      stations]-values)**2, axis=1) / len(stations)
 
         best_point = acceptable_points[np.argmin(scores).ravel()[0]]
